Refactor to move reads and writes into individual domain object models
