+++ app/uk/gov/hmrc/individualsincomeapi/connector/IntegrationFrameworkConnector.scala
@@ -23,8 +23,8 @@ import uk.gov.hmrc.domain.Nino
 import uk.gov.hmrc.http._
 import uk.gov.hmrc.http.logging.Authorization
 import uk.gov.hmrc.individualsincomeapi.domain._
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.{IncomePaye, PayeEntry}
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.sa.{IncomeSa, SaTaxYearEntry}
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.{IFIncomePaye, IFPayeEntry}
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.sa.{IFIncomeSa, IFSaTaxYearEntry}
 import uk.gov.hmrc.play.bootstrap.config.ServicesConfig
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -50,26 +50,26 @@ class IntegrationFrameworkConnector @Inject()(servicesConfig: ServicesConfig, ht
 
   def fetchPayeIncome(nino: Nino, interval: Interval, filter: Option[String])(
     implicit hc: HeaderCarrier,
-    ec: ExecutionContext): Future[Seq[PayeEntry]] = {
+    ec: ExecutionContext): Future[Seq[IFPayeEntry]] = {
 
     val startDate = interval.getStart.toLocalDate
     val endDate = interval.getEnd.toLocalDate
     val payeUrl = s"$serviceUrl/individuals/income/paye/" +
       s"nino/$nino?startDate=$startDate&endDate=$endDate&fields=$filter"
 
-    recover[PayeEntry](http.GET[IncomePaye](payeUrl)(implicitly, header(), ec).map(_.paye))
+    recover[IFPayeEntry](http.GET[IFIncomePaye](payeUrl)(implicitly, header(), ec).map(_.paye))
   }
 
   def fetchSelfAssessmentIncome(nino: Nino, taxYearInterval: TaxYearInterval, filter: Option[String])(
     implicit hc: HeaderCarrier,
-    ec: ExecutionContext): Future[Seq[SaTaxYearEntry]] = {
+    ec: ExecutionContext): Future[Seq[IFSaTaxYearEntry]] = {
 
     val startYear = taxYearInterval.fromTaxYear.endYr
     val endYear = taxYearInterval.toTaxYear.endYr
     val saUrl = s"$serviceUrl/individuals/income/sa/" +
       s"nino/$nino?startYear=$startYear&endYear=$endYear&fields=$filter"
 
-    recover[SaTaxYearEntry](http.GET[IncomeSa](saUrl)(implicitly, header(), ec).map(_.sa))
+    recover[IFSaTaxYearEntry](http.GET[IFIncomeSa](saUrl)(implicitly, header(), ec).map(_.sa))
 
   }
