+++ app/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/paye/IFIncomePaye.scala
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye
+
+import play.api.libs.functional.syntax.{unlift, _}
+import play.api.libs.json.Reads.verifying
+import play.api.libs.json.{Format, JsPath, Reads}
+
+case class IFIncomePaye(paye: Seq[IFPayeEntry])
+
+object IFIncomePaye {
+
+  val minValue = -9999999999.99
+  val maxValue = 9999999999.99
+
+  def isMultipleOfPointZeroOne(value: Double): Boolean = (BigDecimal(value) * 100.0) % 1 == 0
+
+  def isInRange(value: Double): Boolean = value > minValue && value < maxValue
+
+  def paymentAmountValidator(implicit rds: Reads[Double]): Reads[Double] =
+    verifying[Double](value => isInRange(value) && isMultipleOfPointZeroOne(value))
+
+  implicit val incomePayeFormat: Format[IFIncomePaye] = Format(
+    (JsPath \ "paye").read[Seq[IFPayeEntry]].map(value => IFIncomePaye(value)),
+    (JsPath \ "paye").write[Seq[IFPayeEntry]].contramap(value => value.paye)
+  )
+
+}
