+++ app/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/paye/IFPayeEntry.scala
@@ -1,0 +1,120 @@
+/*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye
+
+import play.api.libs.functional.syntax.{unlift, _}
+import play.api.libs.json.{Format, JsPath, Reads}
+import play.api.libs.json.Reads.{maxLength, minLength, pattern, verifying}
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IFIncomePaye._
+
+
+case class IFPayeEntry(
+                        taxCode: Option[String],
+                        paidHoursWorked: Option[String],
+                        taxablePayToDate: Option[Double],
+                        totalTaxToDate: Option[Double],
+                        taxDeductedOrRefunded: Option[Double],
+                        employerPayeRef: Option[String],
+                        paymentDate: Option[String],
+                        taxablePay: Option[Double],
+                        taxYear: Option[String],
+                        monthlyPeriodNumber: Option[String],
+                        weeklyPeriodNumber: Option[String],
+                        payFrequency: Option[String],
+                        dednsFromNetPay: Option[Double],
+                        employeeNics: Option[IFEmployeeNics],
+                        employeePensionContribs: Option[IFEmployeePensionContribs],
+                        benefits: Option[IFBenefits],
+                        parentalBereavement: Option[Double],
+                        studentLoan: Option[IFStudentLoan],
+                        postGradLoan: Option[IFPostGradLoan]
+                    )
+
+object IFPayeEntry {
+
+  val taxCodePattern = "^([1-9][0-9]{0,5}[LMNPTY])|(BR)|(0T)|(NT)|(D[0-8])|([K][1-9][0-9]{0,5})$".r
+  val paidHoursWorkPattern = "^[^ ].{0,34}$".r
+  val employerPayeRefPattern = "^[^ ].{1,9}$".r
+  val paymentDatePattern = ("^(((19|20)([2468][048]|[13579][26]|0[48])|2000)[-]02[-]29|((19|20)[0-9]{2}[-](0[469]|11)" +
+    "[-](0[1-9]|1[0-9]|2[0-9]|30)|(19|20)[0-9]{2}[-](0[13578]|1[02])[-](0[1-9]|[12][0-9]|3[01])|(19|20)[0-9]{2}" +
+    "[-]02[-](0[1-9]|1[0-9]|2[0-8])))$").r
+
+  val payeTaxYearPattern = "^[0-9]{2}\\-[0-9]{2}$".r
+  val monthlyPeriodNumberPattern = "^([1-9]|1[0-2])$".r
+  val weeklyPeriodNumberPattern = "^([1-9]|[1-4][0-9]|5[0-46])$".r
+
+  val payFrequencyValues = Seq("W1", "W2", "W4", "M1", "M3", "M6", "MA", "IO", "IR")
+
+  def isInPayFrequency(implicit rds: Reads[String]): Reads[String] =
+    verifying(value => payFrequencyValues.contains(value))
+
+  implicit val payeEntryFormat: Format[IFPayeEntry] = Format(
+    (
+      (JsPath \ "taxCode").readNullable[String]
+        (minLength[String](2)
+          .keepAnd(maxLength[String](7)
+            .keepAnd(pattern(taxCodePattern, "Invalid Tax Code")))) and
+        (JsPath \ "paidHoursWorked").readNullable[String]
+          (maxLength[String](35)
+            .keepAnd(pattern(paidHoursWorkPattern, "Invalid Paid Hours Work"))) and
+        (JsPath \ "taxablePayToDate").readNullable[Double](paymentAmountValidator) and
+        (JsPath \ "totalTaxToDate").readNullable[Double](paymentAmountValidator) and
+        (JsPath \ "taxDeductedOrRefunded").readNullable[Double](paymentAmountValidator) and
+        (JsPath \ "employerPayeRef").readNullable[String]
+          (maxLength[String](10)
+            .keepAnd(pattern(employerPayeRefPattern, "Invalid employer PAYE reference"))) and
+        (JsPath \ "paymentDate").readNullable[String](pattern(paymentDatePattern, "Invalid Payment Date")) and
+        (JsPath \ "taxablePay").readNullable[Double](paymentAmountValidator) and
+        (JsPath \ "taxYear").readNullable[String](pattern(payeTaxYearPattern, "Invalid Tax Year")) and
+        (JsPath \ "monthlyPeriodNumber").readNullable[String]
+          (pattern(monthlyPeriodNumberPattern, "Invalid Monthly Period Number")
+            .keepAnd(minLength[String](1)).keepAnd(maxLength[String](2))) and
+        (JsPath \ "weeklyPeriodNumber").readNullable[String]
+          (pattern(weeklyPeriodNumberPattern, "Invalid Weekly Period Number")
+            .keepAnd(minLength[String](1)).keepAnd(maxLength[String](2))) and
+        (JsPath \ "payFrequency").readNullable[String](isInPayFrequency) and
+        (JsPath \ "dednsFromNetPay").readNullable[Double](paymentAmountValidator) and
+        (JsPath \ "employeeNICs").readNullable[IFEmployeeNics] and
+        (JsPath \ "employeePensionContribs").readNullable[IFEmployeePensionContribs] and
+        (JsPath \ "benefits").readNullable[IFBenefits] and
+        (JsPath \ "statutoryPayYTD" \ "parentalBereavement").readNullable[Double](paymentAmountValidator) and
+        (JsPath \ "studentLoan").readNullable[IFStudentLoan] and
+        (JsPath \ "postGradLoan").readNullable[IFPostGradLoan]
+      ) (IFPayeEntry.apply _),
+    (
+      (JsPath \ "taxCode").writeNullable[String] and
+        (JsPath \ "paidHoursWorked").writeNullable[String] and
+        (JsPath \ "taxablePayToDate").writeNullable[Double] and
+        (JsPath \ "totalTaxToDate").writeNullable[Double] and
+        (JsPath \ "taxDeductedOrRefunded").writeNullable[Double] and
+        (JsPath \ "employerPayeRef").writeNullable[String] and
+        (JsPath \ "paymentDate").writeNullable[String] and
+        (JsPath \ "taxablePay").writeNullable[Double] and
+        (JsPath \ "taxYear").writeNullable[String] and
+        (JsPath \ "monthlyPeriodNumber").writeNullable[String] and
+        (JsPath \ "weeklyPeriodNumber").writeNullable[String] and
+        (JsPath \ "payFrequency").writeNullable[String] and
+        (JsPath \ "dednsFromNetPay").writeNullable[Double] and
+        (JsPath \ "employeeNICs").writeNullable[IFEmployeeNics] and
+        (JsPath \ "employeePensionContribs").writeNullable[IFEmployeePensionContribs] and
+        (JsPath \ "benefits").writeNullable[IFBenefits] and
+        (JsPath \ "statutoryPayYTD" \ "parentalBereavement").writeNullable[Double] and
+        (JsPath \ "studentLoan").writeNullable[IFStudentLoan] and
+        (JsPath \ "postGradLoan").writeNullable[IFPostGradLoan]
+      ) (unlift(IFPayeEntry.unapply))
+  )
+}
