+++ app/uk/gov/hmrc/individualsincomeapi/domain/integrationframework/sa/IFIncomeSa.scala
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.individualsincomeapi.domain.integrationframework.sa
+
+import play.api.libs.json.Reads.{verifying}
+import play.api.libs.json.{Format, JsPath, Json, Reads}
+
+case class IFIncomeSa(sa: Seq[IFSaTaxYearEntry])
+
+object IFIncomeSa {
+
+  val minValue = -9999999999.99
+  val maxValue = 9999999999.99
+
+  val utrPattern = "^[0-9]{10}$".r
+  val dateStringPattern = ("^(((19|20)([2468][048]|[13579][26]|0[48])|2000)" +
+    "[-]02[-]29|((19|20)[0-9]{2}[-](0[469]|11)[-]" +
+    "(0[1-9]|1[0-9]|2[0-9]|30)|(19|20)[0-9]{2}[-]" +
+    "(0[13578]|1[02])[-](0[1-9]|[12][0-9]|3[01])|(19|20)[0-9]{2}[-]02[-]" +
+    "(0[1-9]|1[0-9]|2[0-8])))$").r
+
+  def isMultipleOfPointZeroOne(value: Double): Boolean = (BigDecimal(value) * 100.0) % 1 == 0
+
+  def isInRange(value: Double): Boolean = value > minValue && value < maxValue
+
+  def paymentAmountValidator(implicit rds: Reads[Double]): Reads[Double] =
+    verifying[Double](value => isInRange(value) && isMultipleOfPointZeroOne(value))
+
+  implicit val incomeSaFormat: Format[IFIncomeSa] = Format(
+    (JsPath \ "sa").read[Seq[IFSaTaxYearEntry]].map(value => IFIncomeSa(value)),
+    (JsPath \ "sa").write[Seq[IFSaTaxYearEntry]].contramap(value => value.sa)
+  )
+
+}
