+++ test/it/uk/gov/hmrc/individualsincomeapi/connectors/IntegrationFrameworkConnectorSpec.scala
@@ -30,8 +30,8 @@ import unit.uk.gov.hmrc.individualsincomeapi.util._
 import utils._
 import play.api.libs.json.Json
 import uk.gov.hmrc.individualsincomeapi.domain.{TaxYear, TaxYearInterval}
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IncomePaye
-import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.sa.IncomeSa
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.paye.IFIncomePaye
+import uk.gov.hmrc.individualsincomeapi.domain.integrationframework.sa.IFIncomeSa
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
@@ -73,12 +73,12 @@ class IntegrationFrameworkConnectorSpec
     wireMockServer.stop()
   }
 
-  val incomePayeNoData = IncomePaye(Seq())
-  val incomePayeSingle = IncomePaye(Seq(createValidPayeEntry()))
-  val incomePayeMulti = IncomePaye(Seq(createValidPayeEntry(), createValidPayeEntry()))
-  val incomeSaNoData = IncomeSa(Seq())
-  val incomeSaSingle = IncomeSa(Seq(createValidSaTaxYearEntry()))
-  val incomeSaMulti = IncomeSa(Seq(createValidSaTaxYearEntry(), createValidSaTaxYearEntry()))
+  val incomePayeNoData = IFIncomePaye(Seq())
+  val incomePayeSingle = IFIncomePaye(Seq(createValidPayeEntry()))
+  val incomePayeMulti = IFIncomePaye(Seq(createValidPayeEntry(), createValidPayeEntry()))
+  val incomeSaNoData = IFIncomeSa(Seq())
+  val incomeSaSingle = IFIncomeSa(Seq(createValidSaTaxYearEntry()))
+  val incomeSaMulti = IFIncomeSa(Seq(createValidSaTaxYearEntry(), createValidSaTaxYearEntry()))
 
   "fetchPaye" should {
